+++ bfabric_scripts/src/bfabric_scripts/cli/login/pat.py
@@ -33,6 +33,9 @@ def cmd_login_pat(
         print("Login aborted.", file=sys.stderr)
         return
     set_default = resolve_set_default(set_default, config_env)
+    if set_default is None:
+        print("Login aborted.", file=sys.stderr)
+        return
 
     if pat is None:
         pat = getpass.getpass("Personal Access Token: ")